requirements.txt
aiofiles>=0.6.0
aiohttp>=3.8.1
colorama>=0.4.4
psutil>=5.9.0
tabulate>=0.8.9
tqdm>=4.62.3
python-socketio>=5.5.0  # For web interface
markdown>=3.3.6  # For documentation generation
pytest>=7.0.0  # For running tests
pyyaml>=6.0  # For configuration handling
